requirements.txt
matplotlib
tensorflow-gpu
scikit-learn